+++ zenith-float-num/tests/mpfr/common.rs
@@ -76,6 +76,41 @@ macro_rules! test_zf_op_no_cc {
     };
 }
 
+macro_rules! test_zf_root {
+    ($eq:literal, $n1:ident, $n:expr, $f1:ident, $p:ident, $rm:ident, $rnd:ident, $op_info:expr, $cc:ident) => {
+        let n3 = ExactNum::nth_root(&($n1), $n, $p, $rm);
+
+        let mut f3 = Float::with_val($p as u32, 1);
+
+        unsafe { mpfr::rootn_ui(f3.as_raw_mut(), ($f1).as_raw(), $n as u64, $rnd) };
+
+        assert_float_close(n3, f3, $p, &format!("{:?}", $op_info), $eq, &mut $cc);
+    };
+}
+
+macro_rules! test_zf_copysign {
+    ($eq:literal, $n1:ident, $n2:ident, $f1:ident, $f2:ident, $p:ident, $rm:ident, $rnd:ident, $op_info:expr, $cc:ident) => {
+        let n3 = ExactNum::copysign(&($n1), &($n2), $p, $rm);
+
+        let mut f3 = Float::with_val($p as u32, 1);
+
+        unsafe { mpfr::copysign(f3.as_raw_mut(), ($f1).as_raw(), ($f2).as_raw(), $rnd) };
+
+        assert_float_close(n3, f3, $p, &format!("{:?}", $op_info), $eq, &mut $cc);
+    };
+}
+
+macro_rules! test_zf_next_after {
+    ($eq:literal, $n1:ident, $n2:ident, $f1:ident, $f2:ident, $p:ident, $rm:ident, $op_info:expr, $cc:ident) => {
+        let n3 = ExactNum::next_after(&($n1), &($n2), $p, $rm);
+
+        let mut f3 = ($f1).clone();
+        unsafe { mpfr::nexttoward(f3.as_raw_mut(), ($f2).as_raw()) };
+
+        assert_float_close(n3, f3, $p, &format!("{:?}", $op_info), $eq, &mut $cc);
+    };
+}
+
 macro_rules! test_zf_fma {
     ($eq:literal, $n1:ident, $n2:ident, $n3:ident, $f1:ident, $f2:ident, $f3:ident, $p:ident, $rm:ident, $rnd:ident, $op_info:expr, $cc:ident) => {
         let n4 = ExactNum::fma(&($n1), &($n2), &($n3), $p, $rm);
@@ -104,6 +139,9 @@ macro_rules! test_zf_const {
     };
 }
 
+pub(crate) use test_zf_root;
+pub(crate) use test_zf_copysign;
+pub(crate) use test_zf_next_after;
 pub(crate) use test_zf_const;
 pub(crate) use test_zf_fma;
 pub(crate) use test_zf_op;